Obstacle Course ~60% complete
I have a decent system down for the robot and he usually gets the ball
on the other side of the folders just fine. However, he just barely
bumps the folders in the process so I'll be adding code to have him put
his arms behind his back so they are not in the way.